+++ History.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Split Strainer class as a factory and a template (#1208) [Thierry Joyal]
+* Remove handling of a nil context in the Strainer class (#1218) [Thierry Joyal]
 
 ## 4.0.3 / 2019-03-12